9.1 管理员未登录，不允许操作。

+++ src/com/JES/action/ManagerAgentManageAction.java
@@ -30,6 +30,10 @@ public class ManagerAgentManageAction extends SuperAction implements
 	 * @return
 	 */
 	public String display() {
+		if (session.getAttribute("managerId") == null) {
+			return "LoginNotYet";
+		}
+		
 		request.setAttribute("agentList", managerService.getAgents());
 		return "agentsDisplay";
 	}
@@ -40,6 +44,10 @@ public class ManagerAgentManageAction extends SuperAction implements
 	 * @return
 	 */
 	public String searchAgent() {
+		if (session.getAttribute("managerId") == null) {
+			return "LoginNotYet";
+		}
+		
 		request.setAttribute("agentList", managerService.searchAgent(
 				request.getParameter("searchType"),
 				request.getParameter("searchValue")));
@@ -47,6 +55,10 @@ public class ManagerAgentManageAction extends SuperAction implements
 	}
 
 	public String toChange() {
+		if (session.getAttribute("managerId") == null) {
+			return "LoginNotYet";
+		}
+		
 		String uid = request.getParameter("uid");
 		session.setAttribute("uid", uid);
 
@@ -54,6 +66,10 @@ public class ManagerAgentManageAction extends SuperAction implements
 	}
 
 	public String toDelete() {
+		if (session.getAttribute("managerId") == null) {
+			return "LoginNotYet";
+		}
+		
 		String uid = request.getParameter("uid");
 		session.setAttribute("uid", uid);
 		session.setAttribute("agent", managerService.getAgentDAO()
@@ -64,12 +80,20 @@ public class ManagerAgentManageAction extends SuperAction implements
 	}
 
 	public String delete() {
+		if (session.getAttribute("managerId") == null) {
+			return "LoginNotYet";
+		}
+		
 		managerService.getAgentDAO().delete(
 				((Agent) session.getAttribute("agent")));
 		return "agentDeleteSuccess";
 	}
 
 	public String change() {
+		if (session.getAttribute("managerId") == null) {
+			return "LoginNotYet";
+		}
+		
 		agent.setUid(request.getParameter("uid"));
 
 		if (managerService.isExistAgent(agent)) {
